src/PPO-BipedalWalker/requirements.txt
torch>=1.9.0
numpy>=1.20.0
gym>=0.21.0
box2d-py>=2.3.8
matplotlib>=3.3.0
tensorboard>=2.7.0
tqdm>=4.60.0
pyyaml>=5.4.0
pytest>=6.0.0
pytest-cov>=2.0.0